+++ examples/d_dmrg.py
@@ -1,6 +1,6 @@
 """Example illustrating the use of DMRG in tenpy.
 
-The example functions in this class do the same as the ones in `a_simple_4_DMRG.py`,
+The example functions in this class do the same as the ones in `toycodes/d_dmrg.py`,
 but make use of the classes defined in tenpy.
 """
 # Copyright 2018 TeNPy Developers
@@ -10,7 +10,7 @@ import numpy as np
 from tenpy.networks.mps import MPS
 from tenpy.models.tf_ising import TFIChain
 from tenpy.algorithms import dmrg
-from a_simple_2_model import TFIModel as SimpleTFIModel
+import tfi_exact
 
 
 def example_DMRG_finite(L, g):
@@ -23,11 +23,10 @@ def example_DMRG_finite(L, g):
     E = np.sum(psi.expectation_value(M.H_bond[1:]))
     print("E = {E:.13f}".format(E=E))
     print("final bond dimensions: ", psi.chi)
-    if L < 20:
-        M2 = SimpleTFIModel(L=L, J=1., g=g, bc='finite')
-        E_ed = M2.exact_finite_gs_energy()
-        print("Exact diagonalization: E = {E:.13f}".format(E=E_ed))
-        print("relative error: ", abs((E - E_ed) / E_ed))
+    if L < 20:  # compare to exact result
+        E_exact = tfi_exact.finite_gs_energy(L, 1., g)
+        print("Exact diagonalization: E = {E:.13f}".format(E=E_exact))
+        print("relative error: ", abs((E - E_exact) / E_exact))
     return E, psi, M
 
 
@@ -42,14 +41,13 @@ def example_DMRG_infinite(g):
     print("E = {E:.13f}".format(E=E))
     print("final bond dimensions: ", psi.chi)
     print("correlation length:", psi.correlation_length())
-    M2 = SimpleTFIModel(L=2, J=1., g=g, bc='infinite')
-    E_ex = M2.exact_infinite_gs_energy()
-    print("Analytic result: E/L = {E:.13f}".format(E=E_ex))
-    print("relative error: ", abs((E - E_ex) / E_ex))
+    # compare to exact result
+    E_exact = tfi_exact.infinite_gs_energy(1., g)
+    print("Analytic result: E/L = {E:.13f}".format(E=E_exact))
+    print("relative error: ", abs((E - E_exact) / E_exact))
     return E, psi, M
 
 
-if __name__ == "__main__":
-    example_DMRG_finite(L=10, g=1.)
-    print("-" * 100)
-    example_DMRG_infinite(g=1.5)
+example_DMRG_finite(L=10, g=1.)
+print("-" * 100)
+example_DMRG_infinite(g=1.5)